requirements.txt
pytorch
sacred
